\name{VcfToTable}
\alias{VcfToTable}
\title{Convert VCF file to data frame}
\usage{
VcfToTable(vcf.file)
}
\arguments{
  \item{vcf.file}{String specifying name of VCF file with .vcf or .vcf.gz format}
}
\value{
Data frame containing data from \code{vcf.file}
}
\description{
Converts VCF file to data frame using output generated by vcfR package.
}
\examples{
vcf.file="myVCF.vcf"
vcf.table=VcfToTable(vcf.file)
}
